pyproject.toml
[project]
name = "eval-actionable-guidance"
version = "0.1.1"
description = "The replication package for the paper 'Can We Trust the Actionable Guidance from Explainable AI Techniques in Defect Prediction?' (SANER 2025)"
authors = [
    { name = "Hansae Ju", email = "sparky@hanyang.ac.kr" }
]
dependencies = [
    "scikit-learn>=1.5.2",
    "lime>=0.2.0.1",
    "seaborn>=0.13.2",
    "mlxtend>=0.23.1",
    "scipy>=1.14.1",
    "natsort>=8.4.0",
    "dice-ml>=0.11",
    "xgboost>=2.1.1",
    "bigml>=9.7.1",
    "ipykernel>=6.29.5",
    "python-dotenv>=1.0.1",
    "tabulate>=0.9.0",
    "cliffs-delta>=1.0.0",
    "statsmodels>=0.14.4",
    "numpy==1.26.4",
    "rpy2>=3.5.17",
]
readme = "README.md"
requires-python = ">= 3.10"

[tool.rye]
virtual = true

[tool.ruff]
target-version = "py310"

[tool.ruff.format]
indent-style = "space"

[tool.ruff.lint]
ignore = ["E741"]

[dependency-groups]
dev = [
    "ruff>=0.8.2",
]
